+++ config.sh
@@ -1,11 +1,10 @@
 #!/bin/bash
 
-# Kubespray version to download. Use "master" for latest master branch.
-KUBESPRAY_VERSION=${KUBESPRAY_VERSION:-2.24.0}
-#KUBESPRAY_VERSION=${KUBESPRAY_VERSION:-master}
+source ./target-scripts/config.sh
 
 # container runtime for preparation node
-docker=${docker:-docker}
+docker=${docker:-podman}
+#docker=${docker:-docker}
 #docker=${docker:-/usr/local/bin/nerdctl}
 
 # Run ansible in container?