+++ download-kubespray-files.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+umask 022
+
 source ./config.sh
 source scripts/common.sh
 source scripts/images.sh
@@ -20,7 +22,11 @@ FILES_DIR=outputs/files
 # kubernetes/cri-tools     : crictl
 # kubernetes/calico/vx.x.x : calico
 # kubernetes/calico        : calicoctl
-# runc/vx.x.x               : runc
+# runc/vx.x.x              : runc
+# cilium-cli/vx.x.x        : cilium-cli
+# gvisor/{ver}/{arch}      : gvisor (sunrc, containerd-shim)
+# scopeo/vx.x.x            : scopeo
+# yq/vx.x.x                : yq
 #
 decide_relative_dir() {
     local url=$1
@@ -32,6 +38,11 @@ decide_relative_dir() {
     rdir=$(echo $rdir | sed "s@.*/crictl-.*.tar.gz@kubernetes/cri-tools@")
     rdir=$(echo $rdir | sed "s@.*/\(v.*\)/calicoctl-.*@kubernetes/calico/\1@")
     rdir=$(echo $rdir | sed "s@.*/\(v.*\)/runc.amd64@runc/\1@")
+    rdir=$(echo $rdir | sed "s@.*/\(v.*\)/cilium-linux-.*@cilium-cli/\1@")
+    rdir=$(echo $rdir | sed "s@.*/\([^/]*\)/\([^/]*\)/runsc@gvisor/\1/\2@")
+    rdir=$(echo $rdir | sed "s@.*/\([^/]*\)/\([^/]*\)/containerd-shim-runsc-v1@gvisor/\1/\2@")
+    rdir=$(echo $rdir | sed "s@.*/\(v[^/]*\)/skopeo-linux-.*@skopeo/\1@")
+    rdir=$(echo $rdir | sed "s@.*/\(v[^/]*\)/yq_linux_*@yq/\1@")
     if [ "$url" != "$rdir" ]; then
         echo $rdir
         return
@@ -47,7 +58,6 @@ decide_relative_dir() {
 
 get_url() {
     url=$1
-    ignore_error=$2
     filename="${url##*/}"
 
     rdir=$(decide_relative_dir $url)
@@ -62,11 +72,12 @@ get_url() {
 
     if [ ! -e $FILES_DIR/$rdir/$filename ]; then
         echo "==> Download $url"
-        curl --location --show-error --fail --output $FILES_DIR/$rdir/$filename $url || {
-            if [ "$ignore_error" != "true" ]; then
-                exit 1
-            fi
-        }
+        for i in {1..3}; do
+            curl --location --show-error --fail --output $FILES_DIR/$rdir/$filename $url && return
+            echo "curl failed. Attempt=$i"
+        done
+        echo "Download failed, exit : $url"
+        exit 1
     else
         echo "==> Skip $url"
     fi
@@ -88,12 +99,7 @@ generate_list() {
     #fi
 }
 
-VENV_DIR=${VENV_DIR:-~/.venv/default}
-if [ ! -e ${VENV_DIR} ]; then
-    echo "No ${VENV_DIR}, abort."
-    exit 1
-fi
-source ${VENV_DIR}/bin/activate
+. ./target-scripts/venv.sh
 
 generate_list
 
@@ -105,14 +111,8 @@ cp ${KUBESPRAY_DIR}/contrib/offline/temp/images.list $IMAGES_DIR/
 # download files
 files=$(cat ${FILES_DIR}/files.list)
 for i in $files; do
-    get_url $i true
-done
-for i in $files; do
-    get_url $i false
+    get_url $i
 done
 
 # download images
-images=$(cat ${IMAGES_DIR}/images.list)
-for i in $images; do
-    get_image $i
-done
+./download-images.sh || exit 1